# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Users & Products API",
    "description": "API for managing users and products",
    "version": "1.0.0"
  },
  "host": "cse341-project1-fphm.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your login token (add 'Bearer ' prefix)"
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "security": [{ "BearerAuth": [] }],
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "List of users" },
          "401": { "description": "Access denied" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": { "type": "string", "example": "John" },
                "lastName": { "type": "string", "example": "Doe" },
                "email": { "type": "string", "example": "test@example.com" },
                "favoriteColor": { "type": "string", "example": "Blue" },
                "birthday": { "type": "string", "format": "date", "example": "1990-01-01" }
              },
              "required": ["firstName", "lastName", "email", "favoriteColor", "birthday"]
            }
          }
        ],
        "responses": {
          "201": { "description": "User created" },
          "401": { "description": "Access denied" },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
...
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all products",
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "List of products" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create a new product",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "example": "Laptop" },
                "description": { "type": "string", "example": "High-performance laptop" },
                "price": { "type": "number", "example": 999.99 },
                "category": { "type": "string", "example": "Electronics" },
                "inStock": { "type": "boolean", "example": true }
              },
              "required": ["name", "description", "price", "category", "inStock"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Product created" },
          "401": { "description": "Access denied" },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
...
  }
}